Moved MainClient to base of package tree. Many modifications to make GUI actually work. Added LevelData for said functionality. Also other things.

+++ Game/src/fog/ethereal/world/SaveableLevel.java
@@ -10,7 +10,7 @@ import javax.xml.bind.annotation.XmlRootElement;
 public class SaveableLevel {
 	private ArrayList<BasicSection> sections;
 	private String name;
-	private double bestTime;
+	private long bestTime;
 	
 	public SaveableLevel(String name) {
 		this.name = name;
@@ -28,11 +28,11 @@ public class SaveableLevel {
 	}
 	
 	@XmlElement(name = "bestTime")
-	public double getBestTime() {
+	public long getBestTime() {
 		return bestTime;
 	}
 	
-	public void setBestTime(double bestTime) {
+	public void setBestTime(long bestTime) {
 		this.bestTime = bestTime;
 	}
